moving Anenx to plugin.h/cpp
