+++ dialects/core/core.thorin
@@ -122,6 +122,7 @@
 /// You can specify the desired behavior in the case of an overflow with the curried argument just in front of the actual argument by providing a thorin::core::Mode as `.Nat`.
 ///
 .ax %core.wrap(add, sub, mul, shl): Π.[s: .Nat][m: .Nat][«2; .Idx s»] -> .Idx s, normalize_wrap;
+
 .lam %core.minus!.(s: .Nat)!(m: .Nat)!(a: .Idx s) -> .Idx s = %core.wrap.sub m (0:(.Idx s), a);
 ///
 /// ### %core.div
